Update README.md
2023-10-26: Info on current beta flow.
Note that there is a newer version than 23.09 available in the list of files above. (23.x beta's).
They are newer, fix a lot of issues, but also not tested thoroughly yet. It can be that the beta is already better and more stable than latest v23.09.
Be aware that the current beta does require some additional information to use correctly. This will all be updated in the help sections in the dashboard. It will take some time create that, after technically this version is working as expected.